requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.1
clarifai==9.9.3
clarifai-grpc==9.8.1
cohere==4.32
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.1
dill==0.3.7
diskcache==5.6.3
fastavro==1.8.2
filelock==3.12.4
FLAML==2.1.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.324
langsmith==0.0.52
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2
nlpcloud==1.1.44
numpy==1.22.0
omegaconf==2.2.3
openai==0.28.1
opencv-python==4.7.0.68
openlm==0.0.5
packaging==23.2
pandas==1.3.5
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.24.4
pyautogen==0.1.13
pycocotools==2.0.6
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.12
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
rich==13.4.2
safetensors==0.4.0
schema==0.7.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlitedict==2.1.0
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
zipp==3.17.0
